requirements.txt
altgraph==0.17.4
packaging==23.2
pefile==2023.2.7
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pywin32-ctypes==0.2.2
PyYAML==6.0.1
